+++ src/components/messages/alert.rs
@@ -1,25 +1,8 @@
 /*
- * Copyright (c) 2024, Stalwart Labs Ltd.
+ * SPDX-FileCopyrightText: 2020 Stalwart Labs Ltd <hello@stalw.art>
  *
- * This file is part of Stalwart Mail Web-based Admin.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of
- * the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Affero General Public License for more details.
- * in the LICENSE file at the top-level directory of this distribution.
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * You can be released from the requirements of the AGPLv3 license by
- * purchasing a commercial license. Please contact licensing@stalw.art
- * for more details.
-*/
+ * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-SEL
+ */
 
 use std::time::Duration;
 
@@ -272,9 +255,7 @@ impl From<http::Error> for Alert {
                 let (title, details) = match error {
                     ManagementApiError::FieldAlreadyExists { field, value } => (
                         "Field already exists".to_string(),
-                        format!(
-                            "Another record exists with value {value:?} in field {field:?}."
-                        ),
+                        format!("Another record exists with value {value:?} in field {field:?}."),
                     ),
                     ManagementApiError::FieldMissing { field } => (
                         "Missing required field".to_string(),